% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_ggcr.R
\name{fit_ggcr}
\alias{fit_ggcr}
\title{fit_ggcr}
\usage{
fit_ggcr(mydat, clo, nu, ni, nt, nc)
}
\arguments{
\item{clo}{number of clones}

\item{nu}{number of updates}

\item{ni}{number of iterations}

\item{nt}{number of thinning iterations}

\item{nc}{number of chains}

\item{mydata}{a matrix of 1's and 2's for non-detections and detection}
}
\value{
This function returns values from the posterior distribution of the parameters of a Gamma-Gompertz model.
}
\description{
This function fits a Gamma-Gompertz model to capture-recapture data using data cloning.
}
\author{
Gilbert Marzolin, Olivier Gimenez
}
\keyword{package}

