% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggcr_model.R
\name{ggcr_model}
\alias{ggcr_model}
\title{ggcr_model: Gamma-Gompertz capture-recapture model}
\usage{
ggcr_model()
}
\description{
This function implements a gamma-gompertz capture-recapture model in the BUGS syntax
}
\author{
Gilbert Marzolin, Olivier Gimenez
}
\keyword{package}

