% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data.R
\name{prep_data}
\alias{prep_data}
\title{prep_data}
\usage{
prep_data(mydata)
}
\arguments{
\item{mydata}{a matrix of 1's and 2's for non-detections and detection}
}
\value{
This function returns a list with components the number of individuals, the number of capture occasions, the occasions of first capture and age.
}
\description{
This function gets number of individuals, number of capture occasions, occasions of first capture and age from a capture-recapture dataset
}
\author{
Gilbert Marzolin, Olivier Gimenez
}
\keyword{package}

